requirements.txt
blinker==1.9.0
python-dotenv==1.2.2
click==8.1.8
Flask==3.1.3
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.6
zipp==3.23.0
pymongo==4.11.3
pytest==8.3.4
Faker==37.1.0
